refactor: centralize image processing
